feat: implement user assignment functionality for letters, including controller, service, and DTO updates.

+++ src/main/java/lk/gov/mohe/adminsystem/user/UserDto.java
@@ -4,13 +4,12 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record UserDto(
-    Long id,
-    String username,
-    String email,
-    String fullName,
-    String phoneNumber,
-    String role,
-    String division,
-    Boolean isActive
-) {
+        Integer id,
+        String username,
+        String email,
+        String fullName,
+        String phoneNumber,
+        String role,
+        String division,
+        Boolean isActive) {
 }
