+++ src/main/java/lk/gov/mohe/adminsystem/user/UserRepository.java
@@ -7,15 +7,15 @@ import org.springframework.lang.NonNull;
 
 import java.util.List;
 
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
     @NonNull
     List<User> findAll();
 
     User findByUsername(String username);
 
     @Query("SELECT u.role.id, COUNT(u) FROM User u WHERE u.role.id IN :roleIds" +
-        " GROUP BY u.role.id")
+            " GROUP BY u.role.id")
     List<Object[]> countUsersByRoleIds(@Param("roleIds") List<Integer> roleIds);
+
+
 }
-
-
